Move Aggregator to a separate proto
