+++ rs/proto/build.rs
@@ -5,6 +5,7 @@ use std::process::Command;
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let proto_file = "./proto/muopdb.proto";
     let admin_proto_file = "./proto/admin.proto";
+    let aggregator_proto_file = "./proto/aggregator.proto";
     let out_dir = PathBuf::from(env::var("OUT_DIR")?);
 
     tonic_build::configure()
@@ -13,7 +14,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .build_server(true)
         .file_descriptor_set_path(out_dir.join("muopdb_descriptor.bin"))
         .out_dir(out_dir)
-        .compile_protos(&[proto_file, admin_proto_file], &["proto"])?;
+        .compile_protos(
+            &[proto_file, admin_proto_file, aggregator_proto_file],
+            &["proto"],
+        )?;
 
     let output = Command::new("cargo")
         .args(&["fmt"])
